Unescape HTML character entities of firstChild
As the comments indicate, we only want the first
text node child of a given element. However, with
our version of BeautifulSoup, calling `prop.children`
blows up.

It is also unclear whether the idea is even correct
about taking the first child only. However, in practice,
it appears that calling `prop.text` grabs the text of
all child nodes and abutts them, which is very ugly.